requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.7
google-cloud-speech==2.21.0
google-cloud-translate==3.12.1
googletrans==4.0.0rc1
langdetect==1.0.9
gtts==2.4.0
pillow==10.1.0
opencv-python==4.8.1.78
# Updated for Python 3.12 compatible wheel
numpy==1.26.4
# Fixed invalid version (0.24.0 does not exist); using latest working
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
requests==2.31.0
httpx==0.25.2
pydub==0.25.1
speechrecognition==3.10.0
pyaudio==0.2.11
geopy==2.4.1
schedule==1.2.2
apscheduler==3.11.0
